Parallelize minidaqs + "Analyze" functionality
Made run_process and info_process in TestHandler a list of processes for each fc7. It now creates a CMSIT_"fc7".xml file instead of CMSIT.xml
file. There's also two terminals for each run_process.

I also added some minor changes in TestValidator and TestHandler so that when it get's a test with "analyze" in the name, it DOESN'T run a test,
and instead just runs TestValidator.

TODO! the image resulting from setResult in TestValidator should pop up in the file tree widget in runwindow, but it doesn't and I'm not totally
sure how to do it.

I also deleted a fair amount of commented-out code because it was bothering me.